chore(storage): split UploadSources into separate files
