+++ packages/swift-google-cloud-storage/Sources/GoogleCloudStorage/UploadSource.swift
@@ -32,165 +32,3 @@ public protocol SeekableUploadSource: UploadSource {
   mutating func seek(to offset: Int64) async throws
 }
 
-/// An upload source that reads from a local file.
-public struct FileSource: SeekableUploadSource {
-  public let fileURL: URL
-  private var offset: Int64 = 0
-
-  public var totalSize: Int64? {
-    do {
-      let values = try fileURL.resourceValues(forKeys: [.fileSizeKey])
-      return values.fileSize.map { Int64($0) }
-    } catch {
-      return nil
-    }
-  }
-
-  public init(fileURL: URL) {
-    self.fileURL = fileURL
-  }
-
-  public mutating func read(maxBytes: Int) async throws -> ByteBuffer? {
-    let handle = try FileHandle(forReadingFrom: fileURL)
-    defer {
-      try? handle.close()
-    }
-    try handle.seek(toOffset: UInt64(offset))
-    guard let data = try handle.read(upToCount: maxBytes), !data.isEmpty else {
-      return nil
-    }
-    offset += Int64(data.count)
-    return ByteBuffer(data)
-  }
-
-  public mutating func seek(to offset: Int64) async throws {
-    guard offset >= 0 else {
-      throw UploadError.internalError("Invalid seek offset: \(offset)")
-    }
-    if let size = totalSize, offset > size {
-      throw UploadError.localSourceTooSmall(localSize: size, gcsOffset: offset)
-    }
-    self.offset = offset
-  }
-}
-
-/// An upload source that wraps in-memory bytes or buffers.
-public struct BytesSource: SeekableUploadSource {
-  public let buffer: ByteBuffer
-  public var totalSize: Int64? {
-    return Int64(buffer.count)
-  }
-  private var offset: Int64 = 0
-
-  public init(buffer: ByteBuffer) {
-    self.buffer = buffer
-  }
-
-  public init(data: Data) {
-    self.buffer = ByteBuffer(data)
-  }
-
-  public mutating func read(maxBytes: Int) async throws -> ByteBuffer? {
-    guard maxBytes > 0, offset < buffer.count else { return nil }
-    let end = min(offset + Int64(maxBytes), Int64(buffer.count))
-    let countToRead = Int(end - offset)
-    let chunk: ByteBuffer
-    switch buffer.storage {
-    case .data(let data):
-      chunk = ByteBuffer(data.subdata(in: Int(offset)..<Int(end)))
-    case .byteBuffer(let nioBuffer):
-      var slice = nioBuffer
-      slice.moveReaderIndex(to: nioBuffer.readerIndex + Int(offset))
-      if let subSlice = slice.readSlice(length: countToRead) {
-        chunk = ByteBuffer(subSlice)
-      } else {
-        chunk = ByteBuffer()
-      }
-    }
-    offset = end
-    return chunk
-  }
-
-  public mutating func seek(to offset: Int64) async throws {
-    guard offset >= 0 else {
-      throw UploadError.internalError("Invalid seek offset: \(offset)")
-    }
-    let size = Int64(buffer.count)
-    guard offset <= size else {
-      throw UploadError.localSourceTooSmall(localSize: size, gcsOffset: offset)
-    }
-    self.offset = offset
-  }
-}
-
-/// An upload source that wraps an arbitrary AsyncSequence of ByteBuffer, Data, or NIOCore.ByteBuffer chunks.
-public struct StreamSource: UploadSource {
-  private final class StateBox: @unchecked Sendable {
-    var nextChunk: () async throws -> NIOCore.ByteBuffer?
-
-    init(nextChunk: @escaping () async throws -> NIOCore.ByteBuffer?) {
-      self.nextChunk = nextChunk
-    }
-  }
-
-  public var totalSize: Int64? { return totalSizeValue }
-  private let totalSizeValue: Int64?
-  private let stateBox: StateBox
-  private var buffer = NIOCore.ByteBuffer()
-
-  public init<S: AsyncSequence & Sendable>(
-    sequence: S,
-    totalSize: Int64? = nil
-  ) where S.Element == ByteBuffer {
-    self.totalSizeValue = totalSize
-    var iterator = sequence.makeAsyncIterator()
-    self.stateBox = StateBox {
-      guard let next = try await iterator.next() else { return nil }
-      return next.byteBuffer
-    }
-  }
-
-  public init<S: AsyncSequence & Sendable>(
-    sequence: S,
-    totalSize: Int64? = nil
-  ) where S.Element == Data {
-    self.totalSizeValue = totalSize
-    var iterator = sequence.makeAsyncIterator()
-    self.stateBox = StateBox {
-      guard let next = try await iterator.next() else { return nil }
-      return ByteBuffer(next).byteBuffer
-    }
-  }
-
-  public init<S: AsyncSequence & Sendable>(
-    sequence: S,
-    totalSize: Int64? = nil
-  ) where S.Element == NIOCore.ByteBuffer {
-    self.totalSizeValue = totalSize
-    var iterator = sequence.makeAsyncIterator()
-    self.stateBox = StateBox {
-      try await iterator.next()
-    }
-  }
-
-  public mutating func read(maxBytes: Int) async throws -> ByteBuffer? {
-    guard maxBytes > 0 else { return nil }
-    while buffer.readableBytes < maxBytes {
-      guard let nextChunk = try await stateBox.nextChunk() else {
-        break
-      }
-      var next = nextChunk
-      buffer.writeBuffer(&next)
-    }
-
-    guard buffer.readableBytes > 0 else {
-      return nil
-    }
-
-    let chunkSize = min(maxBytes, buffer.readableBytes)
-    guard let slice = buffer.readSlice(length: chunkSize) else {
-      return nil
-    }
-    return ByteBuffer(slice)
-  }
-}
